[RegisterArray] split register_array out of struct.py into its own module
Addresses the large-file lint on PR #712: the four register_array definitions
(RegisterArray, register_array, _expand_register_array_naming, _RegisterArrayRef) are
self-contained (only depend on numpy + QuadrantsSyntaxError) and have zero coupling to
struct.py internals, so they belong in their own module like field.py / matrix.py.

  - python/quadrants/lang/register_array.py: new module with the four definitions.
  - python/quadrants/lang/struct.py: drop the inline definitions, import from the new
    module. `__all__` still re-exports the public names so `from quadrants.lang.struct
    import *` (run via `quadrants.lang.__init__`) keeps surfacing them as `qd.register_array`
    and `qd.RegisterArray`.
  - python/quadrants/lang/ast/ast_transformer.py: import `_RegisterArrayRef` directly
    from the new module instead of via struct.py, avoiding a re-export hop.

No behavioural change; 224 tests + 9 xfail across test_register_array.py, test_api.py,
test_py_dataclass.py, test_complex_struct.py, test_struct.py all pass.
