+++ python/quadrants/lang/struct.py
@@ -15,6 +15,12 @@ from quadrants.lang.exception import (
 from quadrants.lang.expr import Expr
 from quadrants.lang.field import Field, ScalarField, SNodeHostAccess
 from quadrants.lang.matrix import Matrix, MatrixType
+from quadrants.lang.register_array import (
+    RegisterArray,
+    _RegisterArrayRef,
+    _expand_register_array_naming,
+    register_array,
+)
 from quadrants.lang.util import (
     cook_dtype,
     in_python_scope,
@@ -595,109 +601,6 @@ class StructField(Field):
         }
         entries["__struct_methods"] = self.struct_methods
         return Struct(entries)
-
-
-class RegisterArray:
-    """Type wrapper for a group of N scalar fields exposed via indexed syntax on a ``@qd.dataclass``.
-
-    See :func:`register_array`. ``register_array(N, dtype)`` annotations on a dataclass cause the
-    StructType to be built with N synthetic scalar fields named ``_{group}0`` .. ``_{group}{N-1}``,
-    while the AST transformer rewrites ``obj.{group}[i]`` to a direct reference to
-    ``obj._{group}{i}`` for python-int / qd.static-resolved indices. PTX/LLVM IR is byte-identical
-    to the hand-rolled named-field equivalent.
-    """
-
-    def __init__(self, count, dtype):
-        if not isinstance(count, int) or count <= 0:
-            raise QuadrantsSyntaxError(f"register_array count must be a positive int, got {count!r}")
-        self.count = count
-        self.dtype = dtype
-
-    def __repr__(self):
-        return f"register_array(count={self.count}, dtype={self.dtype})"
-
-
-def register_array(count, dtype):
-    """Declare a group of ``count`` independently-allocated fields of ``dtype`` on a ``@qd.dataclass``.
-
-    The annotation expands at struct-definition time into ``count`` individually-named scalar
-    members (``_{group}0`` .. ``_{group}{count-1}``). Each member gets its own LLVM ``alloca``,
-    which lets SROA + ``mem2reg`` promote each slot into its own SSA value independently. The
-    motivation is register residency under pressure: ``qd.types.vector(N, dtype)`` collapses
-    into one packed ``alloca`` that the optimiser often spills as a unit when register pressure
-    crosses a threshold; ``register_array`` decomposes the storage up-front so the compiler can
-    keep individual slots in registers and only spill the ones it has to.
-
-    Example::
-
-        @qd.dataclass
-        class Tile:
-            r: qd.register_array(32, qd.f32)   # 32 scalar fields exposed as t.r[0..31]
-
-        t = Tile()
-        t.r[5] = 1.0       # lowers to direct write of synthetic field _r5
-        v = t.r[5]         # same as v = t._r5
-        for k in qd.static(range(32)):
-            t.r[k] = 0.0   # each iter is one AST node, not a 32-way cascade
-
-    For python-int / ``qd.static``-resolved indices, ``t.r[k]`` is rewritten by the AST
-    transformer to the named-field access ``t._r{k}``, producing identical LLVM IR / PTX to
-    a struct declared with N individually-named scalar fields.
-
-    Runtime-int indexing is currently unsupported; use an explicit cascade helper for that
-    case.
-    """
-    return RegisterArray(count, dtype)
-
-
-def _expand_register_array_naming(group_name, index):
-    """Naming convention for the synthetic scalar fields of a ``register_array`` group.
-
-    Public so the AST transformer can mirror this without a circular import.
-    """
-    return f"_{group_name}{index}"
-
-
-class _RegisterArrayRef:
-    """Transient proxy returned by the AST transformer for ``obj.{group}`` where ``group``
-    is a registered ``register_array`` group on the struct type.
-
-    Only valid as the value of a Subscript node: ``obj.{group}[i]``. Resolved by
-    ``ASTTransformer.build_Subscript`` to a direct reference to the synthetic scalar field
-    ``_{group}{i}`` when ``i`` is a python-int / qd.static-resolved integer.
-
-    Used as a not-an-Expr marker; any attempt to use it as a value raises.
-    """
-
-    _qd_is_register_array_ref = True
-
-    def __init__(self, struct, group_name: str, count: int, dtype, naming_fn):
-        self._qd_struct = struct
-        self._qd_group_name = group_name
-        self._qd_count = count
-        self._qd_dtype = dtype
-        self._qd_naming_fn = naming_fn
-
-    def _qd_field_for(self, index: int):
-        if not isinstance(index, (int, np.integer)):
-            raise QuadrantsSyntaxError(
-                f"register_array {self._qd_group_name}[i] requires a python-int index "
-                f"(possibly via qd.static); got runtime index of type {type(index).__name__}"
-            )
-        i = int(index)
-        if i < 0 or i >= self._qd_count:
-            raise QuadrantsSyntaxError(
-                f"register_array index out of bounds: {self._qd_group_name}[{i}] "
-                f"(count={self._qd_count})"
-            )
-        field_name = self._qd_naming_fn(self._qd_group_name, i)
-        return getattr(self._qd_struct, field_name)
-
-    def __repr__(self) -> str:  # pragma: no cover - debug only
-        return (
-            f"<register_array_ref group={self._qd_group_name!r} count={self._qd_count} "
-            f"dtype={self._qd_dtype}>"
-        )
 
 
 class StructType(CompoundType):
